khalil+hana+youssef+wiem+firas+chedli1

+++ Pocker_Back/src/main/java/com/example/pockerplanning/repository/SprintRepository.java
@@ -11,11 +11,11 @@ import java.util.List;
 @Repository
 public interface SprintRepository extends JpaRepository<Sprint, Long> {
     @Query("Select a from Sprint a where a.projet.id=:id_projet")
-    List<Sprint> getprojet_par_sprint(@Param("id_projet") int id_projet);
+    List<Sprint> getprojet_par_sprint(@Param("id_projet") Long id_projet);
 
     @Query("Select a from Sprint a where a.projet.id=:id_projet ")
-    List<Sprint> getprojet_par_cours(@Param("id_projet") int id_projet);
+    List<Sprint> getprojet_par_cours(@Param("id_projet") Long id_projet);
     @Query("Select a from Sprint a where   a.projet.id=:id_projet and a.real_end_date< CURRENT_DATE")
-    List<Sprint> ticket_projet_complet(@Param("id_projet") int id_projet);
+    List<Sprint> ticket_projet_complet(@Param("id_projet") Long id_projet);
 
 }
